==> 01_accessibility_calculation.Rmd <==
---
title: "Workbook 01: Accessibility Calculation"
output: html_notebook
---

```{r, include = FALSE}
library(magrittr)
library(sf)
library(tidyverse)
library(tmap)
library(dplyr)
library(forcats)
library(viridis)
library(patchwork)

# options
tmap_mode("plot")
knitr::opts_chunk$set(echo = FALSE)
```

```{r}
# clear environment
rm(list = ls())
```

# Load Results

```{r}
load("./results/output_workbook_00.RData")
```

# Calculate Accessibility

## Floating Catchment Accessibility

```{r}
#beta <- log(0.1, base=exp(1))/-20
beta <- 0.05

fca_ttm <- ttm %>%
  rename(GeoUID = fromId, doctor_id = toId) %>%
  mutate(doctor_id = as.numeric(doctor_id)) %>% 
  
  # join doctor info
  left_join(doctors %>% 
              st_drop_geometry() %>%
              transmute(doctor_id, doctor_count), by = "doctor_id") %>%
  
  # join population info
  left_join(data_da_2016_poly %>% 
              st_drop_geometry() %>%
              transmute(GeoUID, Population), by = "GeoUID") %>%
  
  # get weighted travel times
  mutate(travel_time_gaus = exp(-beta*travel_time)) %>%
  mutate(fca_weighted_pop = Population * travel_time_gaus) %>%
  
  # group by doctor id and calculate weighted populations and provider-to-population ratios (ppr) for each facility
  group_by(doctor_id) %>%
  mutate(fca_ppr = first(doctor_count) / sum(fca_weighted_pop)) %>%
  ungroup() %>%
  
  mutate(fca_weighted_ppr = fca_ppr * travel_time_gaus)
```

## Balanced FCA

```{r}
bfca_ttm <- fca_ttm %>%
  filter(Population >0) %>%
  group_by(GeoUID) %>%
  mutate(bfca_rowsum = sum(travel_time_gaus),
         bfca_weighted_pop = Population * (travel_time_gaus/bfca_rowsum),
         bfca_ppr_pop_zone = sum(bfca_weighted_pop)) %>%
  ungroup() %>%
  
  # get bfca ppr
  group_by(doctor_id) %>%
  mutate(bfca_ppr_pop_clinic = sum(bfca_weighted_pop),
         bfca_ppr = first(doctor_count) / sum(bfca_weighted_pop),
         bfca_colsum = sum(travel_time_gaus),
         bfca_weighted_ppr = bfca_ppr * (travel_time_gaus/bfca_colsum)) %>%
  ungroup()
```

# Map FCA Results

```{r merge results}
doctors %<>% 
  left_join(fca_ttm %>% 
              transmute(doctor_id, fca_ppr) %>% 
              group_by(doctor_id) %>%
              summarize(fca_ppr = first(fca_ppr)), 
            by = "doctor_id") %>%
  left_join(bfca_ttm %>% 
              transmute(doctor_id, bfca_ppr) %>%
              group_by(doctor_id) %>%
              summarize(bfca_ppr = first(bfca_ppr)),
            by = "doctor_id")

data_da_2016_poly %<>% 
  left_join(fca_ttm %>%
              transmute(GeoUID, fca_weighted_ppr) %>%
              group_by(GeoUID) %>% 
              summarize(fca_access = sum(fca_weighted_ppr)),
            by = "GeoUID") %>%
  left_join(bfca_ttm %>% 
              transmute(GeoUID, bfca_weighted_ppr) %>%
              group_by(GeoUID) %>% 
              summarize(bfca_access = sum(bfca_weighted_ppr)),
            by = "GeoUID")
```

```{r map results}
tm_shape(ontario_poly, bbox = st_bbox(data_da_2016_poly)) + 
  tm_fill("grey90") +
  tm_shape(data_da_2016_poly) + 
  tm_fill(col = "fca_access", palette = "viridis", style = "jenks", title = "FCA Accessibility")  +
  tm_layout(legend.position = c("left","bottom"))

tm_shape(ontario_poly, bbox = st_bbox(data_da_2016_poly)) + 
  tm_fill("grey90") +
  tm_shape(data_da_2016_poly) + 
  tm_fill(col = "bfca_access", palette = "viridis", style = "jenks", title = "BFCA Accessibility")  +
  tm_layout(legend.position = c("left","bottom"))

plot(data_da_2016_poly$fca_access, data_da_2016_poly$bfca_access)

tm_shape(ontario_poly, bbox = st_bbox(data_da_2016_poly)) + 
  tm_fill("grey90") +
  tm_shape(data_da_2016_poly) + 
  tm_fill("grey75") +
  tm_shape(doctors) + 
  tm_bubbles(size = "fca_ppr", 
             col = "fca_ppr", 
             palette = "viridis", 
             style = "jenks", 
             title.size = "FCA PPR", 
             title.col = "") +
  tm_layout(legend.position = c("left","bottom"))

tm_shape(ontario_poly, bbox = st_bbox(data_da_2016_poly)) + 
  tm_fill("grey90") +
  tm_shape(data_da_2016_poly) + 
  tm_fill("grey75") +
  tm_shape(doctors) + 
  tm_bubbles(size = "bfca_ppr", 
             col = "bfca_ppr", 
             palette = "viridis", 
             style = "jenks", 
             title.size = "BFCA PPR", 
             title.col = "") +
  tm_layout(legend.position = c("left","bottom"))

plot(doctors$fca_ppr, doctors$bfca_ppr)
```

# MNL Model

```{r load data}
# use updated travel time matrix
Hamiltontimes <- ttm %>% 
  mutate(travel_time = as.double(travel_time)) %>%
  pivot_wider(id_cols = fromId, names_from = toId, names_prefix = "X", 
              values_from = travel_time, values_fill = 999) %>%
  rename(GeoUID = fromId) %>%
  semi_join(data_da_2016_poly %>% transmute(GeoUID, Households) %>% filter(Households > 0))

Ham_households <- data_da_2016_poly %>% 
  st_drop_geometry() %>% 
  transmute(GeoUID,Data = Households) %>%
  semi_join(Hamiltontimes,"GeoUID")
```

```{r prepare data}
#arrange matrices by ID numbers
Ham_households <- arrange(Ham_households,GeoUID)

doctors <- select(doctors,-1) %>%
  arrange(doctor_id)

Hamiltontimes <- arrange(Hamiltontimes,GeoUID)
```

```{r mnl model step 1}
# Initialize model parameters
alpha = 0.065
omega = 22
H_tot <- sum(Ham_households$Data)
Z_tot <- sum(doctors$doctor_count)
R_bar <- (alpha*H_tot)/(omega*Z_tot)
beta1 <- -0.05
betak2 <- 1
betak3 <- -0.5
GeoUID <- Hamiltontimes$GeoUID
doctor_id <- doctors$doctor_id

# initialize Rj values as equal to average demand to capacity ratio for all DAs
Rj <- data.frame(doctor_id) %>%
  mutate(Ratio = R_bar) 

#set up data frame for trip distributions for each clinic and DA
Tij <- Hamiltontimes 

# set up data frame for calculating exponent term for each clinic and DA
expo <- Hamiltontimes 

# set up data frame to store sums of exponent term
sums <- data.frame(GeoUID) 

# Set up functions for calculating trip distributions
calexpo <- function(expo, Rj)
  {
  expo[,2:181] <- beta1*Hamiltontimes[,2:181]
  
  attractiveness <- betak2*log10(omega*doctors$doctor_count) + betak3*Rj$Ratio
  
  attractiveness <- t(attractiveness)
  
  attractiveness <- attractiveness[col(expo[,2:181])]
  
  attractiveness <- matrix(attractiveness,nrow=nrow(Hamiltontimes),ncol=ncol(Hamiltontimes)-1)

  expo[,2:181] <- expo[,2:181] + attractiveness 
  
  expo[,2:181] <- exp(expo[,2:181])
  
    return(expo)
}

# set up function to calculate trip distributions for each clinic and DA
calTij <- function(expo,sums,Tij) 
  {
  Tij[,2:181] <- ((alpha*Ham_households$Data/sums$Sum)*diag(nrow(Hamiltontimes))) %*%
    as.matrix(expo[,2:181])
  
  return(Tij)
}

#calculate exponent term for all DAs and clinics
expo <- calexpo(expo,Rj) 

# sum each row in exponent matrix
sums$Sum <- apply(expo[,2:181],1,sum) 

# calculate initial trip distribution for all clinics and DAs
Tij <- calTij(expo,sums,Tij) 
```

```{r mnl model step 2 new rj}
#Initialize model parameters
rj <- data.frame(doctor_id)

Rj_new <- data.frame(doctor_id)

convergence <- data.frame(doctor_id)

converged <- data.frame(doctor_id)

epsilon <- 0.01

max_iteration <- 10

iteration <- 1

sums <- data.frame(GeoUID)

# Update Rj values with trip distribution
rj <- rj %>%
  mutate(Ratio = apply(Tij[,2:181],2,sum)/(omega*doctors$doctor_count))

Rj_new <- Rj_new %>%
  mutate(Ratio = (Rj$Ratio + rj$Ratio)/2)

convergence <- convergence %>%
  mutate(Value = abs((Rj_new$Ratio - Rj$Ratio))/Rj$Ratio)

# check if convergence value is less than or equal to epsilon
converged <- converged %>%
  mutate(Value = !(convergence$Value > epsilon)) 

# assign updated ratios to previous ratios
Rj <- Rj_new 

# Iteratively calculate Rj values until they converge
while(!all(converged$Value)) #run loop until all clinics have converged to within epsilon
{
  iteration = iteration + 1
  
  if(iteration == max_iteration) #stop loop if reached max number of iterations
  {
    break
  }
  
  else
  {
    expo <- calexpo(expo, Rj) #calculate exponent terms
    
    sums$Sum <- apply(expo[,2:181],1,sum) #calculate sums of exponent terms
    
    Tij <- calTij(expo,sums,Tij) #calculate trip distribution for this iteration
    
    #update ratios with new trip distribution
    rj <- rj %>%
      mutate(Ratio = apply(Tij[,2:181],2,sum)/(omega*doctors$doctor_count))
    
    Rj_new <- Rj_new %>%
      mutate(Ratio = (Rj$Ratio + rj$Ratio)/2)
    
    convergence <- convergence %>%
      mutate(Value = abs((Rj_new$Ratio - Rj$Ratio))/Rj$Ratio)
  
    converged <- converged %>%
      mutate(Value = !(convergence$Value > epsilon)) #check if ratios have converged
    
    Rj <- Rj_new #assign updated ratios to previous ratios
  }
}

# calculate final exponent terms
expo <- calexpo(expo, Rj) 

# calculate final sums of exponent terms
sums$Sum <- apply(expo[,2:181],1,sum) 

# update trip distribution after final iteration
Tij <- calTij(expo,sums,Tij) 
```

## MNL Accessibilities

```{r}
#Initialize model parameters
ai <- data.frame(GeoUID)

#Calculate accessibilities for each DA
ai <- ai %>%
  mutate(Accessibility = log10(sums$Sum))
```

## Manual Sensitivity Analysis halve betak2

```{r}
#Initialize new model parameters
betak2 <- 0.5

# initialize Rj values as equal to average demand to capacity ratio for all DAs
Rj_k2 <- data.frame(doctor_id) %>%
  mutate(Ratio = R_bar) 

#set up data frame for trip distributions for each clinic and DA
Tij_k2 <- Hamiltontimes 

# set up data frame for calculating exponent term for each clinic and DA
expo_k2 <- Hamiltontimes 

# set up data frame to store sums of exponent term
sums_k2 <- data.frame(GeoUID) 

#calculate exponent term for all DAs and clinics
expo_k2 <- calexpo(expo_k2,Rj_k2) 

# sum each row in exponent matrix
sums_k2$Sum <- apply(expo_k2[,2:181],1,sum) 

# calculate initial trip distribution for all clinics and DAs
Tij_k2 <- calTij(expo_k2,sums_k2,Tij_k2) 

rj_k2 <- data.frame(doctor_id)

Rj_k2_new <- data.frame(doctor_id)

convergence <- data.frame(doctor_id)

converged <- data.frame(doctor_id)

epsilon <- 0.01

max_iteration <- 10

iteration <- 1

sums_k2 <- data.frame(GeoUID)

# Update Rj values with trip distribution
rj_k2 <- rj_k2 %>%
  mutate(Ratio = apply(Tij_k2[,2:181],2,sum)/(omega*doctors$doctor_count))

Rj_k2_new <- Rj_k2_new %>%
  mutate(Ratio = (Rj_k2$Ratio + rj_k2$Ratio)/2)

convergence <- convergence %>%
  mutate(Value = abs((Rj_k2_new$Ratio - Rj_k2$Ratio))/Rj_k2$Ratio)

# check if convergence value is less than or equal to epsilon
converged <- converged %>%
  mutate(Value = !(convergence$Value > epsilon)) 

# assign updated ratios to previous ratios
Rj_k2 <- Rj_k2_new 

# Iteratively calculate Rj values until they converge
while(!all(converged$Value)) #run loop until all clinics have converged to within epsilon
{
  iteration = iteration + 1
  
  if(iteration == max_iteration) #stop loop if reached max number of iterations
  {
    break
  }
  
  else
  {
    expo_k2 <- calexpo(expo_k2, Rj_k2) #calculate exponent terms
    
    sums_k2$Sum <- apply(expo_k2[,2:181],1,sum) #calculate sums of exponent terms
    
    Tij_k2 <- calTij(expo_k2,sums_k2,Tij_k2) #calculate trip distribution for this iteration
    
    #update ratios with new trip distribution
    rj_k2 <- rj_k2 %>%
      mutate(Ratio = apply(Tij_k2[,2:181],2,sum)/(omega*doctors$doctor_count))
    
    Rj_k2_new <- Rj_k2_new %>%
      mutate(Ratio = (Rj_k2$Ratio + rj_k2$Ratio)/2)
    
    convergence <- convergence %>%
      mutate(Value = abs((Rj_k2_new$Ratio - Rj_k2$Ratio))/Rj_k2$Ratio)
  
    converged <- converged %>%
      mutate(Value = !(convergence$Value > epsilon)) #check if ratios have converged
    
    Rj_k2 <- Rj_k2_new #assign updated ratios to previous ratios
  }
}

# calculate final exponent terms
expo_k2 <- calexpo(expo_k2, Rj_k2) 

# calculate final sums of exponent terms
sums_k2$Sum <- apply(expo_k2[,2:181],1,sum) 

# update trip distribution after final iteration
Tij_k2 <- calTij(expo_k2,sums_k2,Tij_k2) 

ai_k2 <- data.frame(GeoUID)

#Calculate new accessibilities for each DA
ai_k2 <- ai_k2 %>%
  mutate(Accessibility = log10(sums_k2$Sum))
```

## Manual Sensitivity Analysis double betak3

```{r}
#Initialize new model parameters
betak2 <- 1
betak3 <- -1

# initialize Rj values as equal to average demand to capacity ratio for all DAs
Rj_k3 <- data.frame(doctor_id) %>%
  mutate(Ratio = R_bar) 

#set up data frame for trip distributions for each clinic and DA
Tij_k3 <- Hamiltontimes 

# set up data frame for calculating exponent term for each clinic and DA
expo_k3 <- Hamiltontimes 

# set up data frame to store sums of exponent term
sums_k3 <- data.frame(GeoUID) 

#calculate exponent term for all DAs and clinics
expo_k3 <- calexpo(expo_k3,Rj_k3) 

# sum each row in exponent matrix
sums_k3$Sum <- apply(expo_k3[,2:181],1,sum) 

# calculate initial trip distribution for all clinics and DAs
Tij_k3 <- calTij(expo_k3,sums_k3,Tij_k3) 

rj_k3 <- data.frame(doctor_id)

Rj_k3_new <- data.frame(doctor_id)

convergence <- data.frame(doctor_id)

converged <- data.frame(doctor_id)

epsilon <- 0.01

max_iteration <- 10

iteration <- 1

sums_k3 <- data.frame(GeoUID)

# Update Rj values with trip distribution
rj_k3 <- rj_k3 %>%
  mutate(Ratio = apply(Tij_k3[,2:181],2,sum)/(omega*doctors$doctor_count))

Rj_k3_new <- Rj_k3_new %>%
  mutate(Ratio = (Rj_k3$Ratio + rj_k3$Ratio)/2)

convergence <- convergence %>%
  mutate(Value = abs((Rj_k3_new$Ratio - Rj_k3$Ratio))/Rj_k3$Ratio)

# check if convergence value is less than or equal to epsilon
converged <- converged %>%
  mutate(Value = !(convergence$Value > epsilon)) 

# assign updated ratios to previous ratios
Rj_k3 <- Rj_k3_new 

# Iteratively calculate Rj values until they converge
while(!all(converged$Value)) #run loop until all clinics have converged to within epsilon
{
  iteration = iteration + 1
  
  if(iteration == max_iteration) #stop loop if reached max number of iterations
  {
    break
  }
  
  else
  {
    expo_k3 <- calexpo(expo_k3, Rj_k3) #calculate exponent terms
    
    sums_k3$Sum <- apply(expo_k3[,2:181],1,sum) #calculate sums of exponent terms
    
    Tij_k3 <- calTij(expo_k3,sums_k3,Tij_k3) #calculate trip distribution for this iteration
    
    #update ratios with new trip distribution
    rj_k3 <- rj_k3 %>%
      mutate(Ratio = apply(Tij_k3[,2:181],2,sum)/(omega*doctors$doctor_count))
    
    Rj_k3_new <- Rj_k3_new %>%
      mutate(Ratio = (Rj_k3$Ratio + rj_k3$Ratio)/2)
    
    convergence <- convergence %>%
      mutate(Value = abs((Rj_k3_new$Ratio - Rj_k3$Ratio))/Rj_k3$Ratio)
  
    converged <- converged %>%
      mutate(Value = !(convergence$Value > epsilon)) #check if ratios have converged
    
    Rj_k3 <- Rj_k3_new #assign updated ratios to previous ratios
  }
}

# calculate final exponent terms
expo_k3 <- calexpo(expo_k3, Rj_k3) 

# calculate final sums of exponent terms
sums_k3$Sum <- apply(expo_k3[,2:181],1,sum) 

# update trip distribution after final iteration
Tij_k3 <- calTij(expo_k3,sums_k3,Tij_k3) 

ai_k3 <- data.frame(GeoUID)

#Calculate new accessibilities for each DA
ai_k3 <- ai_k3 %>%
  mutate(Accessibility = log10(sums_k3$Sum))
```

## Full Sensitivity Analysis 

```{r}
sensitivity_df <- data.frame()
cdr_sensitivity_df <- data.frame()

for (bk2 in 1:10){
  
  for (bk3 in -1:-10){ #-1:-10
    
    #Initialize new model parameters
    betak2 <- bk2/10
    betak3 <- bk3/10
    
    # initialize Rj values as equal to average demand to capacity ratio for all DAs
    Rj_k4 <- data.frame(doctor_id) %>%
      mutate(Ratio = R_bar) 
    
    #set up data frame for trip distributions for each clinic and DA
    Tij_k4 <- Hamiltontimes 
    
    # set up data frame for calculating exponent term for each clinic and DA
    expo_k4 <- Hamiltontimes 
    
    # set up data frame to store sums of exponent term
    sums_k4 <- data.frame(GeoUID) 
    
    #calculate exponent term for all DAs and clinics
    expo_k4 <- calexpo(expo_k4,Rj_k4) 
    
    # sum each row in exponent matrix
    sums_k4$Sum <- apply(expo_k4[,2:181],1,sum) 
    
    # calculate initial trip distribution for all clinics and DAs
    Tij_k4 <- calTij(expo_k4,sums_k4,Tij_k4) 
    
    rj_k4 <- data.frame(doctor_id)
    
    Rj_k4_new <- data.frame(doctor_id)
    
    convergence <- data.frame(doctor_id)
    
    converged <- data.frame(doctor_id)
    
    epsilon <- 0.01
    
    max_iteration <- 10
    
    iteration <- 1
    
    sums_k4 <- data.frame(GeoUID)
    
    # Update Rj values with trip distribution
    rj_k4 <- rj_k4 %>%
      mutate(Ratio = apply(Tij_k4[,2:181],2,sum)/(omega*doctors$doctor_count))
    
    Rj_k4_new <- Rj_k4_new %>%
      mutate(Ratio = (Rj_k4$Ratio + rj_k4$Ratio)/2)
    
    convergence <- convergence %>%
      mutate(Value = abs((Rj_k4_new$Ratio - Rj_k4$Ratio))/Rj_k4$Ratio)
    
    # check if convergence value is less than or equal to epsilon
    converged <- converged %>%
      mutate(Value = !(convergence$Value > epsilon)) 
    
    # assign updated ratios to previous ratios
    Rj_k4 <- Rj_k4_new 
    
    # Iteratively calculate Rj values until they converge
    while(!all(converged$Value)) #run loop until all clinics have converged to within epsilon
    {
      iteration = iteration + 1
      
      if(iteration == max_iteration) #stop loop if reached max number of iterations
      {
        break
      }
      
      else
      {
        expo_k4 <- calexpo(expo_k4, Rj_k4) #calculate exponent terms
        
        sums_k4$Sum <- apply(expo_k4[,2:181],1,sum) #calculate sums of exponent terms
        
        Tij_k4 <- calTij(expo_k4,sums_k4,Tij_k4) #calculate trip distribution for this iteration
        
        #update ratios with new trip distribution
        rj_k4 <- rj_k4 %>%
          mutate(Ratio = apply(Tij_k4[,2:181],2,sum)/(omega*doctors$doctor_count))
        
        Rj_k4_new <- Rj_k4_new %>%
          mutate(Ratio = (Rj_k4$Ratio + rj_k4$Ratio)/2)
        
        convergence <- convergence %>%
          mutate(Value = abs((Rj_k4_new$Ratio - Rj_k4$Ratio))/Rj_k4$Ratio)
      
        converged <- converged %>%
          mutate(Value = !(convergence$Value > epsilon)) #check if ratios have converged
        
        Rj_k4 <- Rj_k4_new #assign updated ratios to previous ratios
      }
    }
    
    # calculate final exponent terms
    expo_k4 <- calexpo(expo_k4, Rj_k4) 
    
    # calculate final sums of exponent terms
    sums_k4$Sum <- apply(expo_k4[,2:181],1,sum) 
    
    # update trip distribution after final iteration
    Tij_k4 <- calTij(expo_k4,sums_k4,Tij_k4) 
    
    ai_k4 <- data.frame(GeoUID)
    
    #Calculate new accessibilities for each DA
    ai_k4 <- ai_k4 %>%
      mutate(access = log10(sums_k4$Sum),
             p_betak2 = betak2,
             p_betak3 = betak3,
             iteration_n = iteration)
    
    sensitivity_df <- rbind(sensitivity_df, ai_k4)
    cdr_sensitivity <- Rj_k4 %>%
      mutate(dcr = Ratio,
             cdr = 1/Ratio,
             p_betak2 = betak2,
             p_betak3 = betak3)
    
    cdr_sensitivity_df <- rbind(cdr_sensitivity_df, cdr_sensitivity)
  }
}
```

# Compare accessibility sensitivity test results

```{r}
# summarize accessibility sensitivities
sensitivity_df_avg <- sensitivity_df %>%
  mutate(p_betak2 = as.factor(p_betak2),
         p_betak3 = as.factor(p_betak3)) %>%
  group_by(p_betak2, p_betak3) %>%
  summarize(access = mean(access), .groups = "drop")

# summarize CDR sensitivities
cdr_sensitivity_df_avg <- cdr_sensitivity_df %>%
  mutate(p_betak2 = as.factor(p_betak2),
         p_betak3 = as.factor(p_betak3)) %>%
  group_by(p_betak2, p_betak3) %>%
  summarize(cdr = mean(cdr),
            dcr = mean(dcr), .groups = "drop")

# join cdr's to accessibilities
sensitivity_df_avg %<>% left_join(cdr_sensitivity_df_avg, by = c("p_betak2", "p_betak3"))

frame_orig <- data.frame(p_betak2 = "1", p_betak3 = "-0.5")
frame_s1 <- data.frame(p_betak2 = "0.5", p_betak3 = "-0.5")
frame_s2 <- data.frame(p_betak2 = "1", p_betak3 = "-1")

# make sensitivity plots
sensitivity_plot <- ggplot(sensitivity_df_avg, aes(x = p_betak2, y = p_betak3)) + 
  geom_raster(aes(fill = access)) + 
  scale_fill_viridis(discrete=FALSE) + 
  scale_y_discrete(limits=rev) +
  labs(x = expression(paste(beta[K+2], " (capacity attractiveness)")),
       y = expression(paste(beta[K+3], " (overcrowding sensitivity)")),
       fill = "Mean \nAccessibility") +
  geom_tile(data=frame_orig, fill = NA, aes(colour = "Original"), size = 1) +
  geom_tile(data=frame_s1, fill = NA, aes(colour = "S1"), size = 1) +
  geom_tile(data=frame_s2, fill = NA, aes(colour = "S2"), size = 1) + 
  scale_color_manual(name = "Scenario", 
                     values =c('Original'='grey40','S1'='orange', "S2" = "blue"), 
                     labels = c('Original','S1', "S2")) + 
  ggtitle("Accessibility")

#sensitivity_plot

cdr_sensitivity_plot <- ggplot(sensitivity_df_avg, aes(x = p_betak2, y = p_betak3)) + 
  geom_raster(aes(fill = cdr)) + 
  scale_fill_viridis(discrete=FALSE, option = "A") + 
  scale_y_discrete(limits=rev) +
  labs(x = expression(paste(beta[K+2], " (capacity attractiveness)")),
       y = expression(paste(beta[K+3], " (overcrowding sensitivity)")),
       fill = "Mean \nCDR") +
  geom_tile(data=frame_orig, fill = NA, aes(colour = "Original"), size = 1) +
  geom_tile(data=frame_s1, fill = NA, aes(colour = "S1"), size = 1) +
  geom_tile(data=frame_s2, fill = NA, aes(colour = "S2"), size = 1) + 
  scale_color_manual(name = "Scenario", 
                     values =c('Original'='grey40','S1'='orange', "S2" = "blue"), 
                     labels = c('Original','S1', "S2")) + 
  ggtitle("Capacity-to-Demand Ratio (CDR)")

#cdr_sensitivity_plot

sensitivity_combined <- cdr_sensitivity_plot + sensitivity_plot + plot_layout(guides = "collect") & theme(legend.position = "bottom")
sensitivity_combined
```

# Join main results back to DA polygons and Doctors

```{r}
#results for beta k2
data_da_2016_poly %<>% 
  left_join(ai_k2 %>% 
              transmute(GeoUID, mnl_access_k2 = Accessibility), 
            by = "GeoUID")

doctors %<>% left_join(Rj_k2_new %>% transmute(doctor_id, mnl_dcr_k2 = Ratio), by = "doctor_id")

#results for beta k3
data_da_2016_poly %<>% 
  left_join(ai_k3 %>% 
              transmute(GeoUID, mnl_access_k3 = Accessibility), 
            by = "GeoUID")

doctors %<>% left_join(Rj_k3_new %>% transmute(doctor_id, mnl_dcr_k3 = Ratio), by = "doctor_id")
```

# Compare Results of original analysis

```{r merge}
data_da_2016_poly %<>% 
  left_join(ai %>% 
              transmute(GeoUID, mnl_access = Accessibility), 
            by = "GeoUID")

doctors %<>% left_join(Rj_new %>% transmute(doctor_id, mnl_dcr = Ratio), by = "doctor_id")
```

```{r plot}
#plot(doctors_ppr$fca_ppr, doctors_ppr$Ratio, main = "cdh ppr vs maria ppr")
#plot(da_fca_access$fca_access, da_fca_access$Accessibility, main = "cdh fca vs maria fca")
#plot(da_fca_access$fca_access, da_fca_access$Value, main = "cdh fca vs maria model")
```

## Map

```{r}
tm_shape(ontario_poly, bbox = st_bbox(data_da_2016_poly)) + 
  tm_fill("grey90") +
  tm_shape(data_da_2016_poly) + 
  tm_fill(col = "fca_access", palette = "viridis", style = "jenks", title = "FCA Accessibility")  +
  tm_layout(legend.position = c("left","bottom"))

tm_shape(ontario_poly, bbox = st_bbox(data_da_2016_poly)) + 
  tm_fill("grey90") +
  tm_shape(data_da_2016_poly) + 
  tm_fill(col = "bfca_access", palette = "viridis", style = "jenks", title = "BFCA Accessibility")  +
  tm_layout(legend.position = c("left","bottom"))

tm_shape(ontario_poly, bbox = st_bbox(data_da_2016_poly)) + 
  tm_fill("grey90") +
  tm_shape(data_da_2016_poly) + 
  tm_fill(col = "mnl_access", palette = "viridis", style = "jenks", title = "MNL Accessibility")  +
  tm_layout(legend.position = c("left","bottom"))

plot(data_da_2016_poly$fca_access, data_da_2016_poly$mnl_access)
plot(data_da_2016_poly$bfca_access, data_da_2016_poly$mnl_access)
```

Sensitivity analysis maps

```{r, eval = FALSE}
tm_shape(ontario_poly, bbox = st_bbox(data_da_2016_poly)) + 
  tm_fill("grey90") +
  tm_shape(data_da_2016_poly) + 
  tm_fill(col = "mnl_access_k2", palette = "viridis", 
          style = "jenks",
          #style = "fixed",breaks = c(0,7.563,7.964,8.219,8.405,8.609), 
          #title = "Beta k3 Sensitivity Test",
          title = expression(paste(beta[K+2], " = 0.5")))  +
  tm_layout(legend.position = c("left","bottom"))

tm_shape(ontario_poly, bbox = st_bbox(data_da_2016_poly)) + 
  tm_fill("grey90") +
  tm_shape(data_da_2016_poly) + 
  tm_fill(col = "mnl_access_k3", palette = "viridis", 
          style = "jenks",
          #style = "fixed",breaks = c(0,7.563,7.964,8.219,8.405,8.609), 
          #title = "Beta k3 Sensitivity Test",
          title = expression(paste(beta[K+3], " = (-1)")))  +
  tm_layout(legend.position = c("left","bottom"))
```

```{r}
MNL_original <- tm_shape(ontario_poly, bbox = st_bbox(data_da_2016_poly)) + 
  tm_fill("grey90") +
  tm_shape(data_da_2016_poly) + 
  tm_fill(col = "mnl_access", palette = "viridis", style = "jenks", title = "MNL \nAccessibility")  +
  tm_layout(title = expression(paste("Original: ", beta[K+2]==1, " & ", beta[K+3]==-0.5)), 
            title.size = .9,
            legend.position = c("left","bottom"), frame = "grey40", frame.lwd = 3)

MNL_k2 <- tm_shape(ontario_poly, bbox = st_bbox(data_da_2016_poly)) + 
  tm_fill("grey90") +
  tm_shape(data_da_2016_poly) + 
  tm_fill(col = "mnl_access_k2", palette = "viridis", 
          style = "jenks",
          #style = "fixed",breaks = c(0,7.563,7.964,8.219,8.405,8.609), 
          #title = "Beta k3 Sensitivity Test",
          title = "MNL \nAccessibility")  +
  tm_layout(title = expression(paste("S1: ", beta[K+2]==0.5, " & ", beta[K+3]==-0.5)), 
            title.size = .9,
            legend.position = c("left","bottom"), frame = "orange", frame.lwd = 3)

MNL_k3 <- tm_shape(ontario_poly, bbox = st_bbox(data_da_2016_poly)) + 
  tm_fill("grey90") +
  tm_shape(data_da_2016_poly) + 
  tm_fill(col = "mnl_access_k3", palette = "viridis", 
          style = "jenks",
          #style = "fixed",breaks = c(0,7.563,7.964,8.219,8.405,8.609), 
          #title = "Beta k3 Sensitivity Test",
          title = "MNL \nAccessibility")  + #expression(paste("S2: ", beta[K+3], " = (-1)"))
  tm_layout(title = expression(paste("S2: ", beta[K+2]==1, " & ", beta[K+3]==-1)), 
            title.size = .9,
            legend.position = c("left","bottom"), frame = "blue", frame.lwd = 3)

#sensitivity_maps <- tmap_arrange(MNL_original, MNL_k2, MNL_k3, nrow = 1, outer.margins = c(0, 0, 0, 0.001))
#tmap_save(sensitivity_maps, filename = "./img/sensitivity_maps.jpg", dpi = 600, width = 9, height = 4)
```


# Save

```{r}
save(data_da_2016_poly, doctors, ontario_poly, ttm, sensitivity_df_avg,
     file = "./results/output_workbook_01.RData", compress = TRUE)
```
